reorganize handlers; update readme
